Refactor security module, added RBAC
