+++ api/api/controllers/security_controller.py
@@ -34,9 +34,128 @@ def login_user(user):
 
 
 @exception_handler
-def get_roles(pretty=False, wait_for_complete=False, offset=0, limit=None, search=None, sort=None):
+def get_users():
+    """Get username of a specified user
+
+    :return: All users data
+    """
+    dapi = DistributedAPI(f=security.get_users,
+                          request_type='local_master',
+                          is_async=False,
+                          logger=logger
+                          )
+    data = raise_if_exc(loop.run_until_complete(dapi.distribute_function()))
+
+    return data, 200
+
+
+@exception_handler
+def get_user(username=None):
+    """Get username of a specified user
+
+    :param username: Username of an user
+    :return: User data
+    """
+    f_kwargs = {'username': username}
+
+    dapi = DistributedAPI(f=security.get_user_id,
+                          f_kwargs=remove_nones_to_dict(f_kwargs),
+                          request_type='local_master',
+                          is_async=False,
+                          logger=logger
+                          )
+    data = raise_if_exc(loop.run_until_complete(dapi.distribute_function()))
+
+    return data, 200
+
+
+@exception_handler
+def check_body(f_kwargs, keys: list = None):
+    """Checks that body is correct
+
+    :param f_kwargs: Body to be checked
+    :param keys: Keys that the body must have only and exclusively
+    :return: 0 -> Correct | str -> Incorrect
+    """
+    if keys is None:
+        keys = ['username', 'password']
+    for key in f_kwargs.keys():
+        if key not in keys:
+            return key
+
+    return 0
+
+
+@exception_handler
+def create_user():
+    """Create a new user
+
+    :return: User data
+    """
+    f_kwargs = {**{}, **connexion.request.get_json()}
+    process = check_body(f_kwargs)
+    if process != 0:
+        raise WazuhError(5005, extra_message='Invalid field found {}'.format(process))
+
+    dapi = DistributedAPI(f=security.create_user,
+                          f_kwargs=remove_nones_to_dict(f_kwargs),
+                          request_type='local_master',
+                          is_async=False,
+                          logger=logger
+                          )
+    data = raise_if_exc(loop.run_until_complete(dapi.distribute_function()))
+
+    return data, 200
+
+
+@exception_handler
+def update_user(username=None):
+    """Modify an existent user
+
+    :param username: Name of the user to be modified
+    :return: User data
+    """
+    f_kwargs = {'username': username, **{}, **connexion.request.get_json()}
+    process = check_body(f_kwargs)
+    if process != 0:
+        raise WazuhError(5005, extra_message='Invalid field found {}'.format(process))
+
+    dapi = DistributedAPI(f=security.update_user,
+                          f_kwargs=remove_nones_to_dict(f_kwargs),
+                          request_type='local_master',
+                          is_async=False,
+                          logger=logger
+                          )
+    data = raise_if_exc(loop.run_until_complete(dapi.distribute_function()))
+
+    return data, 200
+
+
+@exception_handler
+def delete_user(username=None):
+    """Delete an existent user
+
+    :param username: Name of the user to be removed
+    :return: Result of the operation
+    """
+    f_kwargs = {'username': username}
+
+    dapi = DistributedAPI(f=security.delete_user,
+                          f_kwargs=remove_nones_to_dict(f_kwargs),
+                          request_type='local_master',
+                          is_async=False,
+                          logger=logger
+                          )
+    data = raise_if_exc(loop.run_until_complete(dapi.distribute_function()))
+
+    return data, 200
+
+
+@exception_handler
+def get_roles(role_ids=None, pretty=False, wait_for_complete=False, offset=0, limit=None, search=None, sort=None):
     """Returns information from all system roles
 
+    :param role_ids: List of roles ids to be obtained
     :param pretty: Show results in human-readable format
     :param wait_for_complete: Disable timeout response
     :param offset: First item to return.
@@ -46,15 +165,16 @@ def get_roles(pretty=False, wait_for_complete=False, offset=0, limit=None, searc
     :param search: Looks for elements with the specified string
     :return Roles information
     """
+    function = security.get_roles if role_ids is None else security.get_role
     rbac = get_permissions(connexion.request.headers['Authorization'])
-    f_kwargs = {'rbac': rbac, 'offset': offset, 'limit': limit,
+    f_kwargs = {'role_ids': role_ids, 'rbac': rbac, 'offset': offset, 'limit': limit,
                 'sort_by': parse_api_param(sort, 'sort')['fields'] if sort is not None else ['id'],
                 'sort_ascending': True if sort is None or parse_api_param(sort, 'sort')['order'] == 'asc' else False,
                 'search_text': parse_api_param(search, 'search')['value'] if search is not None else None,
                 'complementary_search': parse_api_param(search, 'search')['negation'] if search is not None else None
                 }
 
-    dapi = DistributedAPI(f=security.get_roles,
+    dapi = DistributedAPI(f=function,
                           f_kwargs=remove_nones_to_dict(f_kwargs),
                           request_type='local_master',
                           is_async=False,
@@ -66,31 +186,6 @@ def get_roles(pretty=False, wait_for_complete=False, offset=0, limit=None, searc
     response = Data(data)
 
     return response, 200
-
-
-@exception_handler
-def get_role(role_id, pretty=False, wait_for_complete=False):
-    """Return the information from a system role
-
-    :param role_id: Id of the role to be obtained
-    :param pretty: Show results in human-readable format
-    :param wait_for_complete: Disable timeout response
-    :return Role information
-    """
-    rbac = get_permissions(connexion.request.headers['Authorization'])
-    f_kwargs = {'rbac': rbac, **{'role_id': role_id}}
-
-    dapi = DistributedAPI(f=security.get_role,
-                          f_kwargs=remove_nones_to_dict(f_kwargs),
-                          request_type='local_master',
-                          is_async=False,
-                          wait_for_complete=wait_for_complete,
-                          pretty=pretty,
-                          logger=logger
-                          )
-    data = raise_if_exc(loop.run_until_complete(dapi.distribute_function()))
-
-    return data, 200
 
 
 @exception_handler
@@ -127,44 +222,19 @@ def add_role(pretty=False, wait_for_complete=False):
 
 
 @exception_handler
-def remove_role(role_id, pretty=False, wait_for_complete=False):
-    """Remove one specified role
-
-    :param role_id: Specific role id in the system to be deleted
-    :param pretty: Show results in human-readable format
-    :param wait_for_complete: Disable timeout response
-    :return Operation result
-    """
-    rbac = get_permissions(connexion.request.headers['Authorization'])
-    f_kwargs = {'rbac': rbac, 'role_id': role_id}
-
-    dapi = DistributedAPI(f=security.remove_role,
-                          f_kwargs=remove_nones_to_dict(f_kwargs),
-                          request_type='local_master',
-                          is_async=False,
-                          wait_for_complete=wait_for_complete,
-                          pretty=pretty,
-                          logger=logger
-                          )
-    data = raise_if_exc(loop.run_until_complete(dapi.distribute_function()))
-    response = Data(data)
-
-    return response, 200
-
-
-@exception_handler
-def remove_roles(list_roles=None, pretty=False, wait_for_complete=False):
+def remove_roles(role_ids=None, pretty=False, wait_for_complete=False):
     """Removes a list of roles in the system
 
-    :param list_roles: List of roles ids to be deleted
+    :param role_ids: List of roles ids to be deleted
     :param pretty: Show results in human-readable format
     :param wait_for_complete: Disable timeout response
     :return Two list with deleted roles and not deleted roles
     """
+    function = security.remove_roles if role_ids is None else security.remove_role
     rbac = get_permissions(connexion.request.headers['Authorization'])
-    f_kwargs = {'rbac': rbac, 'list_roles': list_roles}
+    f_kwargs = {'rbac': rbac, 'role_ids': role_ids}
 
-    dapi = DistributedAPI(f=security.remove_roles,
+    dapi = DistributedAPI(f=function,
                           f_kwargs=remove_nones_to_dict(f_kwargs),
                           request_type='local_master',
                           is_async=False,
